--- src/Elders.Cronus.DomainModeling.Tests/When_StringId_is_created.cs
@@ -1,30 +1,0 @@
-﻿//using Machine.Specifications;
-
-//namespace Elders.Cronus
-//{
-//    [Subject("StringTenatnId")]
-//    public class When_StringId_is_created
-//    {
-//        Establish context = () =>
-//        {
-//            id = "123";
-//            NID = "arName";
-//            aggregateName = "arName";
-//            valuePart = id;
-//            urn = new Urn(NID, valuePart);
-//        };
-
-//        Because of = () => result = new StringId(id, aggregateName);
-
-//        It should_have_tenant_as_base_part = () => result.NID.ShouldEqual(urn.NID);
-
-//        It should_have_the_same_value_part = () => result.NSS.ShouldEqual(urn.NSS);
-
-//        static IUrn urn;
-//        static IAggregateRootId result;
-//        static string NID;
-//        static string valuePart;
-//        static string aggregateName;
-//        static string id;
-//    }
-//}
