+++ src/Elders.Cronus.DomainModeling/Cronus.DomainModeling.rn.md
@@ -1,3 +1,6 @@
+#### 6.0.0-beta0012 - 14.04.2020
+* All IDs are now by default URNs instead of haing a property which holds that info.
+
 #### 6.0.0-beta0011 - 10.04.2020
 * Adds DataContract attribute to the Urn class
